feat(???): ³ÂËÉÄÜ¹»´´½¨ºÍÑéÖ¤qq°ó¶¨»á»°

+++ config/chensong.go
@@ -7,7 +7,7 @@ type ChenSongConfig struct {
 	ApiToken       string        `mapstructure:"api_token"`
 	ActivatedQQ    int64         `mapstructure:"activated_qq"`
 	ActivatedGroup int64         `mapstructure:"activated_group"`
-	BindMaxTries   int           `mapstructure:"bind_max_tries"`
+	BindMaxTries   int64         `mapstructure:"bind_max_tries"`
 	BindTimeout    time.Duration `mapstructure:"bind_timeout"`
 	ReceiveToken   string        `mapstructure:"receive_token"`
 }
